+++ BackEnd/com/src/main/kotlin/example/com/models/Task.kt
@@ -1,5 +1,14 @@
 package example.com.models
 
+import org.jetbrains.exposed.dao.IntIdTable
+import org.jetbrains.exposed.sql.Table
+
+object TaskTable : IntIdTable() {
+    val title = varchar("title", 255)
+    val description = text("description")
+    val userId = reference("user_id", UserTable)
+}
+
 data class Task (
     val id: Int,
     val title: String,