+++ BackEnd/com/src/main/kotlin/example/com/models/User.kt
@@ -1,5 +1,13 @@
 package example.com.models
 
+import org.jetbrains.exposed.dao.IntIdTable
+import org.jetbrains.exposed.sql.Table
+
+object UserTable : IntIdTable() {
+    val username = varchar("username", 50).uniqueIndex()
+    val password = varchar("password", 64)
+}
+
 data class User (
     val id: Int,
     val username: String,